requirements.txt
asgiref
atomicwrites
attrs
certifi
cffi
charset-normalizer
colorama
coreapi
coreschema
cryptography
defusedxml
Django
django-filter
django-templated-mail
djangorestframework
djangorestframework-simplejwt
djoser
idna
importlib-metadata
iniconfig
itypes
Jinja2
MarkupSafe
oauthlib
packaging
Pillow
pkg_resources
pluggy
py
pycparser
PyJWT
pyparsing
pytest
pytest-django
pytest-pythonpath
python3-openid
pytz
requests
requests-oauthlib
six
social-auth-app-django
social-auth-core
sqlparse
toml
typing_extensions
uritemplate
urllib3
zipp
